requirements.txt
paho-mqtt==1.5.1
requests==2.25.0
requests-toolbelt==0.9.1
urllib3==1.26.2
Flask
pytz